pyproject.toml
[project]
name = "things_backend"
version = "0.1.0"
description = "Backend for THINGS project"
readme = "README.md"
requires-python = "==3.13.2"
dependencies = [
    "fastapi[standard]==0.135.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.48",
    "bcrypt>=5.0.0",
    "alembic>=1.18.4",
]

[project.optional-dependencies]
sqlite = [
    "aiosqlite>=0.19.0",
]
postgres = [
    "asyncpg>=0.31.0",
]
# Instala todo para desarrollo
dev = [
    "aiosqlite>=0.19.0",
    "asyncpg>=0.31.0",
]
